git-svn-id: http://subversion.sos/svn/sos@13610 c11b8382-0eeb-0310-9acf-d3b3cacd313b

+++ spooler_command.cxx
@@ -1619,7 +1619,7 @@ void Command_processor::execute_http( http::Operation* http_operation, Http_file
             {
                 if( filename_of_path( path ).find( '.' ) == string::npos )      // Kein Punkt: Es muss ein Verzeichnis sein!
                 {
-                    if( !string_ends_with( path, "/" )  &&  isalnum( (uint)*path.rbegin() ) )  // '?' am Ende führt zum erneuten GET mit demselben Pfad
+                    if( !string_ends_with( path, "/" )  &&  isalnum( (uchar)*path.rbegin() ) )  // '?' am Ende führt zum erneuten GET mit demselben Pfad
                     {
                         // (Man könnte hier noch prüfen, ob's wirklich ein Verzeichnis ist.)
                         // Der Browser soll dem Verzeichnisnamen einen Schräger anhängen und das als Basisadresse für weitere Anfragen verwenden.
